requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
Pillow>=8.3.0
albumentations>=1.1.0
tqdm>=4.62.0
tensorboard>=2.7.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
opencv-python>=4.5.0
kagglehub>=0.1.0 